Revert "guns"
This reverts commit a2a154314aed4eb9ec2e7ee7c7f6f7d4df5090c8.
