+++ Assets/Resource/Bullet.prefab
@@ -28,8 +28,8 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 709427693189458192}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -4.0541, y: -2.1421, z: 0}
-  m_LocalScale: {x: 0.48938793, y: 0.12475145, z: 1}
+  m_LocalPosition: {x: -3.79875, y: -1.7844319, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
@@ -97,7 +97,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ba86ba8ffe3a474419b3d413c227e7f4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _layerMasks: 06000000
+  _timeLife: 3
+  _BulletMove: 45
 --- !u!61 &2000076112356654679
 BoxCollider2D:
   m_ObjectHideFlags: 0
