--- Assets/Resource/Circle.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e9882cd8058bac44fad3d0f861cca478
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
