--- Assets/Resource/LanzaGranadas.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 64e54d159dc5a8a4997f0f85423d9de4
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
