--- Assets/Resource/Pistol.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 35632003e2c779d4f98ec63a15e18698
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
