--- Assets/Resource/Square.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2a7ae5be1510e6f47b275e4a302b875c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
