--- Assets/Resource/granade.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: bf4013cd5f3348e4e9598f2805b36764
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
