--- Assets/Resource/trew.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2aa57e6519c30854e9b9e24c9c78ceac
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
