--- Assets/Script/Bullets.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8a68ffac8b1cdf74d8003dcb94f29055
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
