--- Assets/Script/Bullets/Granade.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 812c09383be1bef42918a361ad1a88e8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
