--- Assets/Script/FlyWeight.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1f424ff7cb5a422469334f4477833ae5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
