--- Assets/Script/FlyWeight/GunStats.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5b92b4d9b31d4f14e8fc4ef4e2f88832
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
