--- Assets/Script/InputManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c110da0fefe4d61489caa77b799d1676
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
