--- Assets/Script/ScrpitableObjects.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d973fa5c65a981d4a94eea69914700c0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
