--- Assets/Script/ScrpitableObjects/Granade.asset
@@ -1,21 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5b92b4d9b31d4f14e8fc4ef4e2f88832, type: 3}
-  m_Name: Granade
-  m_EditorClassIdentifier: 
-  _stasts:
-    BulletDamage: 40
-    BulletSpeed: 4
-    BulletLifeTime: 6
-    Cooldown: 6
-    MagSize: 3
-    BulletPrefab: {fileID: 8913299660461384001, guid: bf4013cd5f3348e4e9598f2805b36764, type: 3}
