--- Assets/Script/ScrpitableObjects/Granade.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 42ac7aad1eaa20b47bab4c6d71cd32d7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
