--- Assets/Script/ScrpitableObjects/LanzaGranadas.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: be820e112bf58d04ea030b68f0ac8023
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
