--- Assets/Script/ScrpitableObjects/MachineGun.asset
@@ -1,21 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5b92b4d9b31d4f14e8fc4ef4e2f88832, type: 3}
-  m_Name: MachineGun
-  m_EditorClassIdentifier: 
-  _stasts:
-    BulletDamage: 1
-    BulletSpeed: -20
-    BulletLifeTime: 5
-    Cooldown: 3
-    MagSize: 20
-    BulletPrefab: {fileID: 4823275828548643655, guid: e9882cd8058bac44fad3d0f861cca478, type: 3}
