--- Assets/Script/ScrpitableObjects/MachineGun.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9507ef74a111b1a4fbc25a2b5a347b6d
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
