--- Assets/Script/ScrpitableObjects/Pistol.asset
@@ -1,21 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5b92b4d9b31d4f14e8fc4ef4e2f88832, type: 3}
-  m_Name: Pistol
-  m_EditorClassIdentifier: 
-  _stasts:
-    BulletDamage: 3
-    BulletSpeed: 5
-    BulletLifeTime: 5
-    Cooldown: 20
-    MagSize: 2
-    BulletPrefab: {fileID: 709427693189458192, guid: 93b8952fcf637804f96fa14e5ae8cfa4, type: 3}
