--- Assets/Script/ScrpitableObjects/Pistol.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2826b45a2bb14bc48aca21f283941611
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
