--- Assets/Script/Strategy.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4e9ea63087e83ab42824c84879a7b2e3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
