--- Assets/Script/Strategy/IBullet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 64f0d7d0bcb78ab49865028eedb2da83
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
