--- Assets/Script/Strategy/IGun.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f2c8402d9bb974942a8fb60e9da3378b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
