--- Assets/Script/Strategy/IWeapon.cs
@@ -1,10 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface IWeapon
-{
-    int Damage { get; }
-
-    void Attack();
-}
