--- Assets/Script/Strategy/IWeapon.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 80d5db9a0a08dcd418979d14f70483a7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
