--- Assets/Script/Weapon.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d2cf078aabb29a141ac043d1efb9a550
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
