--- Assets/Script/Weapon/MachineGun.cs
@@ -1,12 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class MachineGun : Gun
-{
-    public override void Attack()
-    {
-        base.Attack();
-
-    }
-}
