--- Assets/Script/Weapon/MachineGun.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b2a958483f2736449a1357cb1d1e2965
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
