--- Assets/Script/Weapon/Pistol.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Pistol : Gun
-{
-    public override void Attack()
-    {
-        base.Attack();
-    }
-}
