--- Assets/Script/Weapon/Pistol.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ed3da32534437be4abd8ed709962ef08
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
