+++ ProjectSettings/TagManager.asset
@@ -15,7 +15,7 @@ TagManager:
   - 
   - Water
   - UI
-  - Enemy
+  - 
   - 
   - 
   -
