--- Zombie_War-rework-.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{1DEDB6D8-1B36-6D98-4E03-9DEB5398B241}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1DEDB6D8-1B36-6D98-4E03-9DEB5398B241}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1DEDB6D8-1B36-6D98-4E03-9DEB5398B241}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1DEDB6D8-1B36-6D98-4E03-9DEB5398B241}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1DEDB6D8-1B36-6D98-4E03-9DEB5398B241}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
